Converting tests to typescript
