+++ x-pack/test/api_integration/services/infraops_graphql_client.js
@@ -4,24 +4,26 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import ApolloClient from 'apollo-boost'; // eslint-disable-line 
 import { format as formatUrl } from 'url';
 import fetch from 'node-fetch';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { ApolloClient } from 'apollo-client';
+import { HttpLink } from 'apollo-link-http';
 
 export function InfraOpsGraphQLProvider({ getService }) {
   const config = getService('config');
   const kbnURL = formatUrl(config.get('servers.kibana'));
 
   return new ApolloClient({
-    uri: `${kbnURL}/api/infra/graphql`,
-    fetch,
-    request: (operation) => {
-      operation.setContext({
-        headers: {
-          'kbn-xsrf': 'xxx'
-        }
-      });
-    }
+    cache: new InMemoryCache(),
+    link: new HttpLink({
+      credentials: 'same-origin',
+      fetch,
+      headers: {
+        'kbn-xsrf': 'xxx',
+      },
+      uri: `${kbnURL}/api/infra/graphql`,
+    }),
   });
 
 }
